ODROID-N1: init: add to handle uevent of tty subsystem
This patch is to handle uevent for tty device in USB bus and create a symbolic
link to the device when its idVendor and idProduct are matched. Especially, in
this patch the known GPS dongle (USB type only) will be handled and the device
of attached hardware will be present as a symbolic link '/dev/ttyGPS'.
Also, set its baudrate when the device is discovered.

The idVendor and idProduct of GPS devices are listed in 'odroid-usbgps.xml'
and located in the root directory.

  <?xml version="1.0" encoding="UTF-8"?>
  <odroid-gps>
      <default device="/dev/ttyACM0" baudrate="9600" />
      <devices>
          <usbdev vid="10c4" pid="ea60" baudrate="9600" />
      </devices>
  </odroid-gps>

Change-Id: I18bd9ff889fae42c985495a7a423a592b55e4e26
Signed-off-by: Dongjin Kim <tobetter@gmail.com>